python/requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
astunparse==1.6.3
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.3
cshogi==0.7.6
cssselect2==0.7.0
distlib==0.3.6
filelock==3.9.0
flatbuffers==23.1.21
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
jsonschema==4.17.3
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
lxml==5.0.1
Markdown==3.4.1
MarkupSafe==2.1.2
msgpack==1.0.4
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pillow==10.2.0
platformdirs==3.0.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.19.3
PyYAML==6.0
ray==2.2.0
reportlab==4.0.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.4
six==1.16.0
svglib==1.5.1
svgwrite==1.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-directml-plugin==0.4.0.dev230202
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.19.0
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.14.0
